Inclusion of Web-Server-and-Client projects
Those projects were created for a code challenge in 2022-05. The challenge was to create a Web load generator to stress test a web api. The challenge provided an online test api, but I created my onw locally, so I could test the performance of my approach without the influence of external factors.